refactor: mts to ts
