+++ package.json
@@ -2,7 +2,7 @@
   "name": "@saucelabs/puppeteer-replay-runner",
   "version": "0.0.0",
   "description": "Sauce Labs Puppeteer Replay Runner",
-  "main": "lib/index.mjs",
+  "main": "lib/index.js",
   "type": "module",
   "files": [
     "lib"
